submodel: Do not depend on aas
Without this change we have a cyclic reference between submodel and aas,
thus we can not build a shared library.

Note: If the submodel has to rely on parts of the aas in the future, the common
parts could be split put into a "aas_shared" or whatever base lib for both, the
aas and the submodel.

Signed-off-by: Tobias Klausmann <tobias.klausmann@lenze.com>