server: Improve library install
With this we can now import the package with find_package(BasyxServer) and
later link with BaSyx::Server.

As an additional benefit we get the search path set and search for the needed
dependencies automatically.

A special case here is asio where we ship our own logic to detect the library.
This could change if asio proivdes its own cmake infrastracture.

Signed-off-by: Tobias Klausmann <tobias.klausmann@lenze.com>