Fixed NullPointerException in OPC UA connector
The `translateBrowsePathToNodeId()` method of didn't check for a
particular failure mode where the client returns `null` if the given
browse path doesn't exist.
A similar check for `getTargets().length == 0` did exist and
has been merged with the new check.
I don't know in which cases `targets` would be an array of zero length and in which cases it is `null`.

Signed-off-by: Moritz Marseu <moritz.marseu@festo.com>
Change-Id: I11bc9f09cd421453f907dd0620a9412930d5394c