ProviderException now accepts Throwable as cause
ProviderException had a constructor which accepted only the type
'Exception' for the 'cause' argument.
This goes against standard Java exception classes which accept the
more general 'Throwable' type.
Since another constructor in ProviderException already accepted
'Throwable' as cause, it can be assumed that this was a mistake and it
is okay to use 'Throwable' in all constructors.

This change makes ProviderException play nice with other exceptions. It
is now possible to take the cause from another exception and create a
new ProviderException from that cause.

Change-Id: Ia71ab40561ee4a97db10f9ac7a735c9a225780c4
Signed-off-by: Moritz Marseu <moritz.marseu@festo.com>